Move admin and loggedIn checks into a module.
